+++ client/src/auth/logout.ts
@@ -8,12 +8,12 @@ import { coreStore } from "../store/core";
 export const Logout = defineComponent({
     // eslint-disable-next-line vue/multi-word-component-names
     name: "Logout",
-    async beforeRouteEnter(_to, _from, next) {
+    async beforeRouteEnter() {
         await http.postJson("/api/logout");
         coreStore.setAuthenticated(false);
         coreStore.setUsername("");
         clearSystems("logging-out");
         socket.disconnect();
-        next({ path: "/auth/login" });
+        return { path: "/auth/login" };
     },
 });