+++ client/src/invitation.ts
@@ -1,19 +1,19 @@
 import { defineComponent } from "vue";
-import type { NavigationGuardNext, RouteLocationNormalized } from "vue-router";
+import type { RouteLocationNormalized } from "vue-router";
 
 import { http } from "./core/http";
 
 export default defineComponent({
-    async beforeRouteEnter(to: RouteLocationNormalized, _from: RouteLocationNormalized, next: NavigationGuardNext) {
+    async beforeRouteEnter(to: RouteLocationNormalized) {
         const response = await http.postJson("/api/invite", {
             code: to.params.code,
         });
         if (response.ok) {
             const data = (await response.json()) as { sessionUrl: string };
-            next({ path: data.sessionUrl });
+            return { path: data.sessionUrl };
         } else {
             console.error("Invitation code could not be redeemed");
-            next({ path: "/dashboard" });
+            return { path: "/dashboard" };
         }
     },
 });